Extract types to different files
