+++ src/common/types/common.ts
@@ -1,0 +1,12 @@
+export interface Header {
+  value: string;
+  style?: string;
+  title?: string;
+  sortable?: boolean;
+  includeHeader?: boolean;
+  centerHeader?: boolean;
+}
+
+export type DateObject = {
+  "@ts": string;
+};
